# Conflicts:
#	src/main/java/com/example/sprint2be/controller/ProductController.java

+++ src/main/java/com/example/sprint2be/controller/UserRestController.java
@@ -63,31 +63,35 @@ public class UserRestController {
 
     @PostMapping("/login")
     public ResponseEntity<?> authenticateUser(@Valid @RequestBody Login loginRequest) throws AuthenticationException {
-        Authentication authentication = authManager.authenticate(
-                new UsernamePasswordAuthenticationToken(
-                        loginRequest.getUsername(),
-                        loginRequest.getPassword()
-                )
-        );
-
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-        String token = jwtProvider.generatingJwt(authentication);
-        UserPrincipal userPrincipal = (UserPrincipal) authentication.getPrincipal();
-        JwtResponse response = new JwtResponse(
-                token,
-                userPrincipal.getUsername(),
-                userPrincipal.getEmail(),
-                userPrincipal.getAvatar(),
-                userPrincipal.getAuthorities()
-        );
-        //        quan
         User user = userService.findByUsername(loginRequest.getUsername());
-        Date today = new Date(System.currentTimeMillis());
-        SimpleDateFormat timeFormat = new SimpleDateFormat("hh:mm:ss dd/MM/yyyy");
-        String s = timeFormat.format(today.getTime());
-        user.setSignInRecent(s);
-        userService.updateUser(user);
-        return ResponseEntity.ok(response);
+        if (user.getEnabled().equals("true")) {
+            Authentication authentication = authManager.authenticate(
+                    new UsernamePasswordAuthenticationToken(
+                            loginRequest.getUsername(),
+                            loginRequest.getPassword()
+                    )
+            );
+
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+            String token = jwtProvider.generatingJwt(authentication);
+            UserPrincipal userPrincipal = (UserPrincipal) authentication.getPrincipal();
+            JwtResponse response = new JwtResponse(
+                    userPrincipal.getId().toString(),
+                    token,
+                    userPrincipal.getUsername(),
+                    userPrincipal.getEmail(),
+                    userPrincipal.getAvatar(),
+                    userPrincipal.getAuthorities()
+            );
+            //        quan
+            Date today = new Date(System.currentTimeMillis());
+            SimpleDateFormat timeFormat = new SimpleDateFormat("hh:mm:ss dd/MM/yyyy");
+            String s = timeFormat.format(today.getTime());
+            user.setSignInRecent(s);
+            userService.updateUser(user);
+            return ResponseEntity.ok(response);
+        }
+        return new ResponseEntity<>(HttpStatus.CONFLICT);
     }
 
     @GetMapping("/user")
@@ -137,7 +141,7 @@ public class UserRestController {
                     "Bạn vừa yêu cầu phục hồi lại mật khẩu đăng nhập của tài khoản: " + recoverPassword.getUsername() + "\n" +
                             "Mã Xác Nhận: " + confirmCode + "\n" +
                             "Xin sử dụng đường dẫn sau đây và nhập vào mã xác nhận trong email này để có thể thực hiệc việc đặt lại mật khẩu mới:" +
-                            " http://localhost:4200/recover-password?confirmCode=" + confirmCode +"&username="+ recoverPassword.getUsername() +"\n"+
+                            " http://localhost:4200/recover-password?username" + recoverPassword.getUsername() +"\n"+
                             "Nếu đây không phải là bạn, xin hãy đăng nhập và thực hiện thay đổi mật khẩu hiện tại.");
             return new ResponseEntity<>(HttpStatus.OK);
         } else {
@@ -146,9 +150,9 @@ public class UserRestController {
     }
 
     @PostMapping("/check-code/{confirmCode}")
-    public ResponseEntity<?> checkConfirmCode(@PathVariable String confirmCode) {
+    public ResponseEntity<?> checkConfirmCode(@PathVariable String confirmCode, @RequestParam String username) {
         Optional<RecoverPassword> checkExist = recoverPasswordService.loadByConfirmCode(confirmCode);
-        if (checkExist.isPresent()) {
+        if ((checkExist.isPresent()) && (checkExist.get().getUsername().equals(username))) {
             User user = userService.findByUsername(checkExist.get().getUsername());
             UserDto userDto = userService.convertToUserDto(user);
             recoverPasswordService.delete(checkExist.get());
@@ -157,9 +161,9 @@ public class UserRestController {
         return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
-    @PostMapping("/change-password/{id}")
-    public ResponseEntity<?> changePassword(@PathVariable Integer id, @RequestParam String password) {
-        if (Boolean.TRUE.equals(userService.changePassword(id, password))) {
+    @PostMapping("/change-password/{username}")
+    public ResponseEntity<?> changePassword(@PathVariable String username, @RequestParam String password) {
+        if (Boolean.TRUE.equals(userService.changePassword(username, password))) {
             return new ResponseEntity<>(HttpStatus.OK);
         } else return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
